Show opponent average ELO per match to one decimal place.
Use fractional averages on the opponents screen (e.g. +2.7) instead of
truncating to whole numbers.

+++ app/src/main/java/com/rpsonline/app/domain/WeeklyOpponents.kt
@@ -11,8 +11,8 @@ data class WeeklyOpponentRow(
     val matchCount: Int,
     val lastPlayedAt: Long,
 ) {
-    fun avgMyEloDeltaPerMatch(): Int =
-        if (matchCount > 0) weeklyEloDelta / matchCount else 0
+    fun avgMyEloDeltaPerMatch(): Double =
+        if (matchCount > 0) weeklyEloDelta.toDouble() / matchCount else 0.0
 }
 
 private data class OpponentWeekMatchSlice(
